Refactor for testability
